Fix: Refactor tool pages for Next.js metadata and client components
This addresses issues where:
- "use client" was not at the top of the file.
- Metadata was defined in Client Components.

I refactored a significant number of tool pages under app/tools/ to:
- Separate client-side logic into new `client.tsx` files.
- Ensure `client.tsx` files start with "use client".
- Make `page.tsx` files Server Components responsible for metadata.
- Add or preserve metadata (title, description) for refactored tool pages.
- Remove developer comments from refactored files.

This is a partial refactoring of all tools, as you requested.
The following tools were among those refactored:
ascii-to-base64, base64-decode, html-to-base64, image-to-base64,
base64-to-image, cmyk-to-pantone, color-desaturation,
css-border-generator, crc-32-hash-generator, and many others.
